+++ changelog.md
@@ -8,6 +8,8 @@ A change like this was requested for a while...
 #### Improvements
 
 - `hata.ext` directory now handles sub-file imports better.
+- Add `weak` parameter to `bind`.
+- Add `weak_cache_size` parameter to `weak_cache_size`.
 
 ##### ext.slash